added test and authentication, and refactored stuff
