+++ controllers/access.js
@@ -35,7 +35,6 @@ export async function grantAccess(req, reply, next) {
     }
 }
 
-
 export async function revokeAccess(req, reply, next) {
     try {
         const value = req.body.value;
@@ -55,6 +54,36 @@ export async function revokeAccess(req, reply, next) {
 
         return reply.json({
             message: "REVOKED"
+        })
+    } catch (err) {
+        return next(err)
+    }
+}
+
+export async function getUsersWithAccess(req, reply, next) { //TODO: test this.
+    try {
+        const value = req.body.value;
+        const directoryId = value.directoryId;
+
+        const users = await prisma.user_directory_access.findMany({
+            where: {
+                directory_id: directoryId
+            },
+            select: {
+                user_rights: true,
+                user: {
+                    select: {
+                        id: true,
+                        email: true,
+                    }
+                }
+            }
+        })
+        if (!users) throw new APIError();
+
+        return reply.json({
+            message: "OK",
+            users: users
         })
     } catch (err) {
         return next(err)
