+++ validators/bookmark.js
@@ -1,21 +1,31 @@
 import Joi from "joi";
 import { objectValidator, singleValidator } from "./basic_validators.js";
+import { ValidationError } from "../util/error.js";
 
 const createBookmarkValidation = Joi.object({
-    link: Joi.string().uri().required(), //TODO: is URI is like URL? 
-    owner_id: Joi.number().required(),
-    directory_id: Joi.number().required(),
-    type: Joi.string().required(), //TODO:  has to be in the enum. 
+    link: Joi.string().required().custom(validateUrl, 'url validation'),
+    ownerId: Joi.number().required(),
+    directoryId: Joi.number().required(),
+    type: Joi.string().valid('img', 'link', 'etc').required(),
     favorite: Joi.boolean().required()
 })
 
 const updateBookmarkValidation = Joi.object({
     id: Joi.number().required(),
-    link: Joi.string().uri().required(), //TODO: is URI is like URL? 
-    directory_id: Joi.number().required(),
-    type: Joi.string().required(), //TODO:  has to be in the enum. 
+    link: Joi.string().required().custom(validateUrl, 'url validation'),
+    directoryId: Joi.number().required(),
+    type: Joi.string().valid('img', 'link', 'etc').required(),
     favorite: Joi.boolean().required()
 })
+
+function validateUrl  (value, helpers) {
+    try {
+        new URL(value); // Attempt to create a URL object
+        return value;    // URL is valid
+    } catch (error) {
+        return helpers.error('string.uri', { value }); // URL is invalid
+    }
+};
 
 const idValidation = Joi.number().required();
 
@@ -23,19 +33,7 @@ export function createBookmarkDataValidator(req, reply, next) {
     try {
         const value = objectValidator(createBookmarkValidation, req.body);
 
-        req.body.value = value;
-        next()
-    } catch (err) {
-        return next(err);
-    }
-}
-
-export function userIdValidator(req, reply, next) {
-    try {
-        const id = req.body.userId; // this is extracted after auth. 
-        const value = singleValidator(idValidation, id);
-
-        req.body.value = value;
+        req.body = { value: value }
         next()
     } catch (err) {
         return next(err);
@@ -47,7 +45,7 @@ export function bookmarkIdValidator(req, reply, next) {
         const id = req.params.id;
         const value = singleValidator(idValidation, id);
 
-        req.body.value = value;
+        req.body = { value: { bookmarkId: value } }
         next()
     } catch (err) {
         return next(err);
@@ -59,7 +57,7 @@ export function tagIdValidator(req, reply, next) {
         const id = req.params.tagId;
         const value = singleValidator(idValidation, id);
 
-        req.body.value = value;
+        req.body = { value: { tagId: value } }
         next()
     } catch (err) {
         return next(err);
@@ -68,15 +66,14 @@ export function tagIdValidator(req, reply, next) {
 
 export function updateBookmarkDataValidator(req, reply, next) {
     try {
-        const updateList = req.params.ids;
+        const changesList = req.body.changes;
+        if (!changesList) throw new ValidationError();
 
-        const list = [];
-        updateList.forEach(element => {
-            const value = singleValidator(updateBookmarkValidation, element);
-            list.push(value);
+        changesList.forEach(change => {
+            objectValidator(updateBookmarkValidation, change);
         });
 
-        req.body.value = list;
+        req.body = { value: { changes: changesList } }
         next()
     } catch (err) {
         return next(err);
@@ -85,15 +82,16 @@ export function updateBookmarkDataValidator(req, reply, next) {
 
 export function deleteBookmarkDataValidator(req, reply, next) {
     try {
-        const updateList = req.params.ids;
+        const updateList = req.query.ids;
+        if (!updateList) throw new ValidationError();
 
-        const list = [];
+        const idList = [];
         updateList.forEach(id => {
             const value = singleValidator(idValidation, id);
-            list.push(value);
+            idList.push(value);
         });
 
-        req.body.value = list;
+        req.body = { value: { ids: idList } }
         next()
     } catch (err) {
         return next(err);
